+++ src/pages/products/[productId]/index.tsx
@@ -1,11 +1,11 @@
 import { useSetAtom } from "jotai";
 import { useRouter } from "next/router";
 import { useState } from "react";
-import { addToCartAtom } from "~/components/cartStore";
-import Header from "~/components/products/Header";
+import { addToCartAtom } from "@/lib/products/cartStore";
+import Header from "@/lib/products/Header";
 import { Button } from "@/ui/button";
 import { Input } from "@/ui/input";
-import Layout from "~/components/ui/Layout";
+import Layout from "@/lib/components/Layout";
 import {
   Select,
   SelectContent,
@@ -16,7 +16,7 @@ import {
 import { useToast } from "@/ui/use-toast";
 import { type RouterInputs, api } from "~/utils/api";
 import Image from "next/image";
-import { DisclaimerText } from "~/components/products/DisclaimerText";
+import { DisclaimerText } from "@/lib/products/DisclaimerText";
 
 type Size = RouterInputs["order"]["add"]["products"][number]["size"];
