+++ src/pages/products/cart/[[...index]].tsx
@@ -5,10 +5,10 @@ import {
   clearCartAtom,
   removeFromCartAtom,
   updateCartItemQuantityAtom,
-} from "~/components/cartStore";
-import Header from "~/components/products/Header";
+} from "@/lib/products/cartStore";
+import Header from "@/lib/products/Header";
 import { Input } from "@/ui/input";
-import Layout from "~/components/ui/Layout";
+import Layout from "@/lib/components/Layout";
 import { Button } from "@/ui/button";
 import { DialogContent, DialogTitle, DialogTrigger } from "@/ui/dialog";
 import dynamic from "next/dynamic";
@@ -16,11 +16,11 @@ import { api } from "~/utils/api";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { type AddFormOrder, addFormOrderSchema } from "~/schemas/order";
-import { FieldValidation } from "~/components/ui/FieldValidation";
+import { FieldValidation } from "@/lib/components/FieldValidation";
 import { useToast } from "@/ui/use-toast";
 import { useState } from "react";
 import Image from "next/image";
-import { DisclaimerText } from "~/components/products/DisclaimerText";
+import { DisclaimerText } from "@/lib/products/DisclaimerText";
 
 // Prevent Nextjs hydration warning
 const ClientSideDialog = dynamic(
